pyproject.toml
[tool.poetry]
name = "36-kuda-ugodno-backend-api"
version = "0.3.0"
description = "API сервис для веб-сайта 'Куда Угодно'"
authors = [
    "Никита Шидогубов",
    "Анна Зылёва",
    "Евгений Кудряшов",
    "Андрей Гусельников",
]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.4"
djangorestframework = "^3.16.0"
pillow = "^11.3.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.1"
gunicorn = "^23.0.0"
django-phonenumber-field = "^8.1.0"
phonenumbers = "^9.0.8"
celery = "^5.5.3"
redis = "^6.2.0"
django-celery-beat = "^2.8.1"
drf-spectacular = "^0.28.0"
django-cors-headers = "^4.7.0"
drf-yasg = "^1.21.10"
dj-rest-auth = "^7.0.1"
django-allauth = "^65.9.0"
social-auth-app-django = "^5.5.1"
djangorestframework-simplejwt = "^5.5.0"
django-filter = "^25.1"
django-autocomplete-light = "^3.12.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
black = "^25.1.0"
ruff = "^0.12.4"

[tool.poetry.group.test.dependencies]
factory-boy = "^3.3.3"
pytest-django = "^4.11.1"
pytest = "^8.4.1"

[tool.ruff]
line-length = 119
exclude = [
    ".git",
    ".gitignore",
    "__pycache__",
    "__init__.py",
    "manage.py",
    "poetry.lock",
    "pyproject.toml",
    "migrations",
    "tests.py",
]
target-version = "py312"
fix = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint]
ignore = ["E501", "W293"]
select = ["E", "F", "I", "B", "C", "A", "N", "UP"]
mccabe.max-complexity = 12

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
